pyproject.toml
[project]
name = "ai-trading-system"
version = "0.1.0"
description = "Multi-agent AI trading system"
authors = [
    {name = "AI Developer", email = "ai.dev@example.com"}
]
dependencies = [
    "pandas>=2.1.0",
    "numpy>=1.24.0",
    "aiohttp>=3.8.0",
    "asyncio-throttle>=1.0.2",
    "python-dotenv>=1.0.0",
    "pydantic>=2.0.0",
    "pydantic-settings>=2.0.0",
    "openai>=1.0.0",
    "groq>=0.4.0",
    "yfinance>=0.2.18",
    "requests>=2.31.0",
    "redis>=4.5.0",
    "fastapi>=0.100.0",
    "uvicorn>=0.23.0",
    "sqlalchemy>=2.0.0",
    "alembic>=1.11.0",
    "ta-lib>=0.4.28",
    "pandas-ta>=0.3.14b",
    "plotly>=5.15.0",
    "matplotlib>=3.7.0",
    "seaborn>=0.12.0",
    "scikit-learn>=1.3.0",
    "backtrader>=1.9.78",
    "zipline-reloaded>=3.0.0",
    "quantstats",
    "empyrical-reloaded",
    "schedule>=1.2.0",
    "APScheduler>=3.10.0",
    "structlog>=23.1.0",
    "rich>=13.4.0",
    "typer>=0.9.0",
    "langchain>=0.1.0",
    "langgraph>=0.0.30",
    "pytest-asyncio>=1.1.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.uv]
dev-dependencies = [
    "jupyter>=1.0.0",
    "ipython>=8.14.0",
    "notebook>=6.5.0",
    "pytest>=7.4.0",
    "pytest-asyncio>=0.21.0",
    "pytest-cov>=4.1.0",
    "pytest-mock>=3.12.0",
    "black>=23.7.0",
    "isort>=5.12.0",
    "flake8>=6.0.0",
    "mypy>=1.5.0",
    "pre-commit>=3.3.0",
]

[tool.hatch.build.targets.wheel]
packages = ["src"]

[tool.pytest.ini_options]
pythonpath = [
  "."
]
testpaths = [
  "tests"
]
markers = [
  "unit: Unit tests for individual components",
  "integration: Integration tests for component interactions",
  "e2e: End-to-end tests for complete workflows",
  "performance: Performance and profiling tests",
  "validation: Validation and quality tests",
  "comparison: Comparison tests"
]
